Revert "Type cleanup for serializarion"
This reverts commit 82b0c10b89f1b962bed1f06b15d662b11ae1398a.

+++ src/g_blake/blake_sbar.cpp
@@ -183,7 +183,7 @@ void BlakeStatusBar::DrawStatusBar()
 
 	// Draw bottom information
 	FString health;
-	health.Format("%3d", (int) players[0].health);
+	health.Format("%3d", players[0].health);
 	DrawString(HealthFont, health, 128, 162, false);
 
 	FString score;
